Qualidade: magic numbers nomeados, coinType consistente, NIP-06 accountIndex (5.4/5.5/6.3)
WalletInit.kt: removido .toInt() redundante em network.coinType (já é Int
desde antes — WalletRestore.kt nunca teve esse .toInt() extra, os dois
agora idênticos nesse ponto).

SegwitSigner.buildP2PKHScript(): opcodes 0x76/0xa9/0x14/0x88/0xac viram
constantes nomeadas (OP_DUP, OP_HASH160, PUSH_20_BYTES, OP_EQUALVERIFY,
OP_CHECKSIG) em vez de literais só com comentário ao lado.

ChangePlanner: dust limit 546L viraQ constante DUST_LIMIT_P2WPKH pública,
default do parâmetro passou a referenciá-la em vez de repetir o literal.

FeeEstimator: pesos de vbyte (68/31/10) viram VBYTES_PER_INPUT/
VBYTES_PER_OUTPUT/VBYTES_OVERHEAD.

NostrKeys.deriveFromSeed(): ganhou parâmetro accountIndex: Int = 0
(default preserva comportamento atual) — path NIP-06 tinha o índice de
conta fixo em 0, impedindo múltiplas identidades Nostr da mesma seed.

Item 5.3 (toHex()/hexToBytes() acessíveis globalmente) já estava OK,
confirmado sem necessidade de mudança (HexExtensions.kt, pacote
com.pokewallet.crypto).

Itens 6.1 (BITCHAT_GEOHASH configurável) e 6.2 (timeout/retry com
backoff no NostrRelayClient) do review NÃO alterados nesta sessão —
avaliados como baixa prioridade (assim classificados pelo próprio
review) e desproporcionais ao esforço/risco frente ao que já foi
entregue: geohash já é uma constante nomeada trivial de editar;
timeoutMs já é parâmetro configurável do caller, faltando só
retry-with-backoff, que exigiria redesenhar o ciclo de vida dos
WebSockets do broadcast BitChat (caminho testado e funcionando) por um
ganho marginal.

+++ src/main/java/com/pokewallet/crypto/WalletInit.kt
@@ -61,7 +61,7 @@ object WalletInit {
             seed,
             intArrayOf(
                 KeyDerivation.hardened(purpose),
-                KeyDerivation.hardened(network.coinType.toInt()),
+                KeyDerivation.hardened(network.coinType),
                 KeyDerivation.hardened(0)
             )
         )
